moved block store config to storage folder
