+++ storage/block/config.go
@@ -18,7 +18,7 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-package blockstore
+package block
 
 const (
 	DefaultAvailabilityWindow = 8192
